Fix directory name display when MARK_DIRS is set
The expansion pattern ${NAME##*/} is used to retrieve the last part of a path.
Unfortunately, if NAME represents the path to a directory and the MARK_DIRS
option is set, this will end up deleting everything in NAME, as a forward slash
is the last character in the string, resulting in nothing being displayed for
the directory name.

To work around this, the trailing slash is deleted if it's there, then added
back in if the option is set.